feat: creating charts component
